refactor GuessNumberConsoleController
