+++ games/guessnumber-console/src/main/java/ch/zhaw/gameconsoleapp/guessnumber/console/controller/GuessNumberConsoleController.java
@@ -1,28 +1,28 @@
-package ch.zhaw.gameconsoleapp.guessnumber.console.service;
+package ch.zhaw.gameconsoleapp.guessnumber.console.controller;
 
 import ch.zhaw.gameconsoleapp.guessnumber.core.components.CoreGameLogic;
 import ch.zhaw.gameconsoleapp.guessnumber.core.components.MessageProvider;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Controller;
 
 import java.util.InputMismatchException;
 import java.util.Scanner;
 
 /**
- * GuessNumberConsoleService
+ * GuessNumberConsoleController
  * Start class for the Guess Number Game.
  *
  * @author created by Urs Albisser, on 2019-12-30
  * @version 1.0.1
  */
-@Service
-public class GuessNumberConsoleService {
+@Controller
+public class GuessNumberConsoleController {
 
 
 	// == constants ==
-	private static final Logger log = LoggerFactory.getLogger(GuessNumberConsoleService.class);
+	private static final Logger log = LoggerFactory.getLogger(GuessNumberConsoleController.class);
 
 
 	// == fields ==
@@ -32,13 +32,13 @@ public class GuessNumberConsoleService {
 
 	// == constructors ==
 	/**
-	 * GuessNumberConsoleService()
+	 * GuessNumberConsoleController()
 	 * Autowired Constructor
 	 * @param coreGameLogic CoreGameLogic instance to be initialized
 	 * @param messageProvider MessageProvider instance to be initialized
 	 */
 	@Autowired
-	public GuessNumberConsoleService(CoreGameLogic coreGameLogic, MessageProvider messageProvider) {
+	public GuessNumberConsoleController(CoreGameLogic coreGameLogic, MessageProvider messageProvider) {
 		this.coreGameLogic = coreGameLogic;
 		this.messageProvider = messageProvider;
 	}
@@ -51,7 +51,7 @@ public class GuessNumberConsoleService {
 	 */
 	public void startGame() {
 
-		log.info("GuessNumberConsoleService.startGame() called. Game starts up...");
+		log.info("GuessNumberConsoleController.startGame() called. Game starts up...");
 
 		Scanner scanner = new Scanner(System.in);
